refactor: move util/helper functions into their own folders and split concerns
